requirements.txt
python-dotenv>=0.10.4
streamlit>=1.22.0
pymupdf>=1.22.2
langchain>=0.0.160
openai>=0.27.6
tiktoken>=0.3.3
pinecone-client>=2.2.1
rich>=13.3.5